+++ packages/apps/ability-erc20-approval/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 3.1.4 (2025-10-08)
+
+### 🧱 Updated Dependencies
+
+- Updated ability-sdk to 2.3.1
+- Updated app-sdk to 2.2.3
+
+## 3.1.3 (2025-10-06)
+
+### 🧱 Updated Dependencies
+
+- Updated ability-sdk to 2.3.0
+- Updated app-sdk to 2.2.2
+
 ## 3.1.2 (2025-09-25)
 
 ### 🧱 Updated Dependencies